requirements.txt
langchain
langchain-community
tiktoken
chromadb
langgraph
tavily-python
python-dotenv
google-generativeai
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
langchain-mistralai
wikipedia
langchain-huggingface
google-search-results
faiss-cpu
sentence-transformers
